fix: improve android script; add capacitor-set-version dependency

+++ android.sh
@@ -1,3 +1,9 @@
+if [ -z "$1" ]
+  then
+    echo "No argument supplied"
+    exit
+fi
+
 node node_modules/vite/bin/vite.js build --mode android
 rm -rf ./bundle
 mkdir bundle
@@ -5,6 +11,11 @@ npx cap add android
 npx cap copy android
 rm -rf ../dist
 cordova-res android --skip-config --copy
+# Set build number
+build=$(cat buildnumber)
+newbuild=$(($build+1))
+node_modules/capacitor-set-version/bin/run set:android -v $1 -b $newbuild
+
 cd android
 # To enable persistent cookies
 echo "package com.wwwallet.app;
@@ -28,3 +39,4 @@ rm -rf ./android
 cd bundle
 jarsigner -sigalg SHA256withRSA -digestalg SHA-256 -keystore ~/keys/wwwallet.jks -signedjar wwwallet.aab wwwallet-unsigned.aab wwwallet
 rm wwwallet-unsigned.aab
+echo $newbuild > ../buildnumber
